Added image delete option

+++ client/src/screens/admin/ProductListScreen.js
@@ -6,8 +6,13 @@ import { LinkContainer } from 'react-router-bootstrap'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faEdit } from '@fortawesome/free-regular-svg-icons'
 import { faTrash } from '@fortawesome/free-solid-svg-icons'
+import { useAlert } from 'react-alert'
+import Loader from '../../components/Loader'
+import Message from '../../components/Message'
+import { DELETE_PRODUCT_RESET } from '../../constants/productConstants'
 
 const ProductListScreen = ({ history }) => {
+  const alert = useAlert()
   const dispatch = useDispatch()
   //check logged in user
   const userLogIn = useSelector((state) => state.userLogIn)
@@ -16,15 +21,31 @@ const ProductListScreen = ({ history }) => {
   const productList = useSelector((state) => state.productList)
   const { loading, error, products } = productList
 
+  const productDelete = useSelector((state) => state.productDelete)
+  const {
+    loading: loadingDelete,
+    error: errorDelete,
+    success: successDelete,
+  } = productDelete
+
   useEffect(() => {
     if (userInfo && userInfo.role !== 'admin') {
       history.push('/')
-    } else {
-      dispatch(listProduct())
     }
   }, [dispatch, userInfo, history])
+
+  useEffect(() => {
+    dispatch(listProduct())
+    if (successDelete) {
+      alert.success('Product Deleted')
+      dispatch({ type: DELETE_PRODUCT_RESET })
+    }
+  }, [dispatch, successDelete, alert])
+
   const deleteHandler = (id) => {
-    dispatch(deleteProduct(id))
+    if (window.confirm('Are You Sure?')) {
+      dispatch(deleteProduct(id))
+    }
   }
   return (
     <>
@@ -32,59 +53,65 @@ const ProductListScreen = ({ history }) => {
         <Col>
           <h1>Products</h1>
         </Col>
+        <Col>{loading || (loadingDelete && <Loader></Loader>)}</Col>
       </Row>
-      <Table striped bordered hover responsive className='table-sm'>
-        <thead>
-          <tr>
-            <th>SL No</th>
-            <th>IMAGE</th>
-            <th>NAME</th>
-            <th>PRICE</th>
-            <th>CATEGORY</th>
-            <th>ADDONS</th>
-            <th>ACTION</th>
-          </tr>
-        </thead>
-        <tbody>
-          {products.map((product, index) => (
-            <tr key={product._id}>
-              <td>{index + 1}</td>
-              <td>
-                <Image
-                  src={product.image.url}
-                  rounded
-                  style={{ width: '40px' }}
-                />
-              </td>
-              <td>{product.title}</td>
-              <td>${product.price}</td>
-              <td>{product.category.name}</td>
-              <td>
-                {product.addon.map((a) => (
-                  <>
-                    <span>{a.name}</span>
-                    <br />
-                  </>
-                ))}
-              </td>
-              <td>
-                <LinkContainer to={`/admin/product/${product._id}/edit`}>
-                  <Button variant='dark' className='btn-sm'>
-                    <FontAwesomeIcon icon={faEdit} />
-                  </Button>
-                </LinkContainer>
-                <Button
-                  variant='danger'
-                  className='btn-sm'
-                  onClick={() => deleteHandler(product._id)}
-                >
-                  <FontAwesomeIcon icon={faTrash} />
-                </Button>
-              </td>
+      <Row>
+        {error && <Message variant='danger'>{error}</Message>}
+        {errorDelete && <Message variant='danger'>{errorDelete}</Message>}
+
+        <Table striped bordered hover responsive className='table-sm'>
+          <thead>
+            <tr>
+              <th>SL No</th>
+              <th>IMAGE</th>
+              <th>NAME</th>
+              <th>PRICE</th>
+              <th>CATEGORY</th>
+              <th>ADDONS</th>
+              <th>ACTION</th>
             </tr>
-          ))}
-        </tbody>
-      </Table>
+          </thead>
+          <tbody>
+            {products.map((product, index) => (
+              <tr key={product._id}>
+                <td>{index + 1}</td>
+                <td>
+                  <Image
+                    src={product.image.url}
+                    rounded
+                    style={{ width: '40px' }}
+                  />
+                </td>
+                <td>{product.title}</td>
+                <td>${product.price}</td>
+                <td>{product.category.name}</td>
+                <td>
+                  {product.addon.map((a) => (
+                    <>
+                      <span>{a.name}</span>
+                      <br />
+                    </>
+                  ))}
+                </td>
+                <td>
+                  <LinkContainer to={`/admin/product/${product._id}/edit`}>
+                    <Button variant='dark' className='btn-sm'>
+                      <FontAwesomeIcon icon={faEdit} />
+                    </Button>
+                  </LinkContainer>
+                  <Button
+                    variant='danger'
+                    className='btn-sm'
+                    onClick={() => deleteHandler(product._id)}
+                  >
+                    <FontAwesomeIcon icon={faTrash} />
+                  </Button>
+                </td>
+              </tr>
+            ))}
+          </tbody>
+        </Table>
+      </Row>
     </>
   )
 }
